diagnosis_inference/workflows/tag_rag_proper/requirements.txt
openai>=1.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
tqdm>=4.65.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
